backend/requirements.txt
# Backend dependencies for ICS Security Monitoring System

# FastAPI framework
fastapi>=0.70.0
uvicorn>=0.15.0
pydantic>=1.8.2

# Database
sqlalchemy>=1.4.27
psycopg2-binary>=2.9.2
alembic>=1.7.5

# Caching and messaging
redis>=4.0.2
celery>=5.2.0

# WebSockets
websockets>=10.1

# ICS protocol handling
pymodbus>=3.0.0,<4.0.0
scapy>=2.4.5

# ML/AI libraries
numpy>=1.20.3
pandas>=1.3.4
scikit-learn>=1.0.1
tensorflow>=2.8.0
joblib>=1.1.0

# Network scanning
python-nmap>=0.7.1

# Utilities
python-dotenv>=0.19.2
requests>=2.26.0
PyYAML>=6.0

# Additional dependencies for ARFF data processing
lxml>=4.6.0 